compromiso 5 ok

+++ backend/PCM.Application/Features/Com5EstrategiaDigital/Queries/GetCom5EstrategiaDigitalQuery.cs
@@ -2,6 +2,6 @@ namespace PCM.Application.Features.Com5EstrategiaDigital.Queries;
 
 public class GetCom5EstrategiaDigitalQuery
 {
-    public int CompromisoId { get; set; }
+    public long CompromisoId { get; set; }
     public Guid EntidadId { get; set; }
 }
